run.py
import os
from dotenv import load_dotenv

load_dotenv()

from app import create_app
from app.extensions import socketio

env = os.environ.get('FLASK_ENV', 'development')
app = create_app(env)

if __name__ == '__main__':
    socketio.run(app, debug=True, port=5050, allow_unsafe_werkzeug=True)
