requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-WTF==1.2.2
Flask-Cors==5.0.1

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Background tasks
celery[redis]==5.4.0
redis==5.2.1

# WebSocket
eventlet==0.37.0

# Auth
Werkzeug==3.1.3
email-validator==2.2.0
PyJWT==2.10.1

# S3
boto3==1.35.86

# PDF processing
pdf2image==1.17.0
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
gunicorn==23.0.0
